+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './api.js';
+export * from './interceptor.js';
+export * from './service.js';
+export * from './be-object.js';