+++ CHANGELOG.md
@@ -1,3 +1,57 @@
+# [4.0.0-beta.5](https://github.com/atlasconsulting/bedita-sdk-js/compare/v4.0.0-beta.4...v4.0.0-beta.5) (2026-07-06)
+
+
+### Bug Fixes
+
+* defult value for  MapIncludedInterceptor.#config ([7efda1f](https://github.com/atlasconsulting/bedita-sdk-js/commit/7efda1fc14ae601665e10caa1de6c69efec7f96a)), closes [MapIncludedInterceptor.#config](https://github.com/MapIncludedInterceptor./issues/config)
+
+
+### Features
+
+* flatten relationships meta ([5218783](https://github.com/atlasconsulting/bedita-sdk-js/commit/521878320c16cd0ffe7a61a91596278ba5ff1436))
+* improve types for flatten objects ([8af6b6d](https://github.com/atlasconsulting/bedita-sdk-js/commit/8af6b6deebd62ecf41146b8b361fdda847b79f42))
+
+# [4.0.0-beta.4](https://github.com/atlasconsulting/bedita-sdk-js/compare/v4.0.0-beta.3...v4.0.0-beta.4) (2026-06-08)
+
+
+### Bug Fixes
+
+* axios security issues ([9d995f9](https://github.com/atlasconsulting/bedita-sdk-js/commit/9d995f9cd015c24e8761567f1bde8b8327532daa))
+
+# [4.0.0-beta.3](https://github.com/atlasconsulting/bedita-sdk-js/compare/v4.0.0-beta.2...v4.0.0-beta.3) (2026-03-10)
+
+
+### Bug Fixes
+
+* update import paths to include .js extension ([335b86f](https://github.com/atlasconsulting/bedita-sdk-js/commit/335b86fd1e3b096558fa1c83efbce7934ff173c8))
+
+# [4.0.0-beta.2](https://github.com/atlasconsulting/bedita-sdk-js/compare/v4.0.0-beta.1...v4.0.0-beta.2) (2026-03-10)
+
+
+### Bug Fixes
+
+* export from .js files ([18cd4df](https://github.com/atlasconsulting/bedita-sdk-js/commit/18cd4df7b5c7c36b589505ea1a6d7186a52d4a87))
+
+# [4.0.0-beta.1](https://github.com/atlasconsulting/bedita-sdk-js/compare/v3.3.2...v4.0.0-beta.1) (2026-03-06)
+
+
+### Bug Fixes
+
+* security axios upgrade ([c58c2e3](https://github.com/atlasconsulting/bedita-sdk-js/commit/c58c2e3aaaa9eaf6a165d4be49cb518d07beeea7))
+* wrong import ([9bc1322](https://github.com/atlasconsulting/bedita-sdk-js/commit/9bc1322713558f6bd150cf527ccfb2d6cf99deb9))
+
+
+### Features
+
+* add flatten option to MapIncludedInterceptor ([af8efcf](https://github.com/atlasconsulting/bedita-sdk-js/commit/af8efcfe2fae0284a523f53f706f42156265247d))
+* add helpers to flatten, exapand JSON api resource ([d7bf084](https://github.com/atlasconsulting/bedita-sdk-js/commit/d7bf084a271b9c829ac759e044361165cc30961b))
+* upgrade deps ([4753e6b](https://github.com/atlasconsulting/bedita-sdk-js/commit/4753e6b771d14314983cc5ef541f4ea67bfb986c))
+
+
+### BREAKING CHANGES
+
+* refresh all deps
+
 ## [3.3.3](https://github.com/atlasconsulting/bedita-sdk-js/compare/v3.3.2...v3.3.3) (2026-06-08)